+++ scripts/sp-login.sh
@@ -1,0 +1,23 @@
+#!/bin/bash
+
+# logout first
+echo "Logging out..."
+az logout || true
+
+# fetch the values
+AZ_APPID=$(jq -r .appId ./secrets/sp.json)
+AZ_PASSWORD=$(jq -r .password ./secrets/sp.json)
+AZ_TENANT=$(jq -r .tenant ./secrets/sp.json)
+AZ_ENVID=$(jq -r .envId ./secrets/sp.json)
+AZ_SUBSCRIPTION=$(jq -r .subscriptionId ./secrets/sp.json)
+
+echo "Logging in as ADMIN service principal..."
+az login --service-principal \
+    --username $AZ_APPID \
+    --password $AZ_PASSWORD \
+    --tenant $AZ_TENANT
+az account set --subscription $AZ_SUBSCRIPTION -o table
+
+# check result
+az account show -o table
+