+++ CHANGELOG.md
@@ -4,4 +4,6 @@
   - Started new repo based on experience witth `mkol5222/automagic-june-2025`
   - Start with Terraform `.gitignore`
   - DevContainer/CodeSpaces configuration to include Terraform CLI and Azure CLI
+  - ADMIN and READER service principals creation, test and login
+